SAFRAN-496: Add possibility to specify several indexes.
In an Entity model, the PHYSICAL_UNIQUE expression can be used to
specify the indexes to generate in the database through scaffolding.
This adds the possibility to specify *several* such indexes, separated
by the '|' character.

The reverse transformation, from a logical database to an Entity model
also supports this syntax.

Change-Id: I19e2ae2f988152fa410264d557a8afb4411b69cf
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>