Fix linkedRequirement view edit and create actions
Since we moved to EEF 2 all EEF 1.5 properties View contribution where
removed. This broke the linkedRequirement view edit and create actions
which open a EEF 1.5 wizard on the fly, which is something EEF 2 cannot
do easily.

As a quick fix we decided to re-activate the EEF 1.5 provider for
Requirement to make linkRequirement view edit and create actions work
again.

Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>