Update to EMF Compare 3.3.11
This version is required to avoid getting an NPE when opening the
comparison editor. See
https://support.jira.obeo.fr/browse/SAFRAN-847?focusedCommentId=3459577&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-3459577

Upgrading this version raised a problem regarding the version of guava.
EMF Compare 3.3.11 requires a guava [22.0.0,27.0.0). So this commit
refactors all plugins that require guava to use package import instead
of plugin dependencies. It allows the product to have two guava versions
working side by side (21.0.0 and 27.1.0).

It then raises a problem with fr.obeo.dsl.designer.gen.html which
incorrectly declares its dependencies constraints. We are currently
discussing the problem with the product team to find a solution. In the
meantime we will remove that feature from the product.