[SAFRAN-843] Fix random problem with Entity wizard
Note : The author of this fix is Laurent Redor (laurent.redor@obeo.fr)

Sometimes, the tool "Namespace" in the "Entities Namespaces Hierarchy" diagram has not the expected effect. The "Namespace" instance is created but does not appear in the diagram. This is a side effect of a more global problem. Indeed, the representation created by the wizard is sometimes wrongly initialized. This is caused by a save that can be in progress during the addition of the new representation in the resource (the CDOTransactionChangeRecorder adapter is not added on the new representation). This save is automatically launched by the CollabSaveSessionWhenNoDialectEditorsListener because of execution of the previous recording command (to activate viewpoint).

This commit adds a "join" on the save job to be sure it is finished before creating the representations.