[SAFRAN-1064]Restore opposite Reference upon library update
When a project uses a library and we update this library, some semantic
elements may disappear as a result of the library update.

For instance this is the case if a bidirectional relation is defined
between an entity from the project and a entity from the library: upon
library update, the EReference link 'referencedType' from the Reference
in the project entity can be restored to point to the corresponding
Entity of the library, but its opposite Reference does not exist and
thus is not set as 'oppositeOf' to the Reference.

Now upon a library update, we have a look at whether a Reference has an
opposite Reference, in which case we try to re-created that Reference
after the library has been updated. And we set up the 'referencedType'
and 'oppositeOf' EReferences such that the bidirectional References
between both entities look like what they were before.

Note that to make these restored References appear on the Entity
diagram, a close/open of the whole Session is required.